Renamed saddle-vesdl strategy to saddle-sdl-balance-with-lock
