+++ src/strategies/index.ts
@@ -251,7 +251,7 @@ import * as oceanDAOBrightID from './ocean-dao-brightid';
 import * as saddleFinance from './saddle-finance';
 import * as saddleFinanceV2 from './saddle-finance-v2';
 import * as vesdlVotingPower from './saddle-vesdl-voting-power';
-import * as vesdl from './saddle-vesdl';
+import * as sdlBalanceWithLock from './saddle-sdl-balance-with-lock';
 import * as lydiaGovVault from './lydia-gov-vault';
 import * as xkawaFarm from './xkawa-farm';
 import * as darkforestScore from './darkforest-score';
@@ -578,7 +578,7 @@ const strategies = {
   'ocean-dao-brightid': oceanDAOBrightID,
   'saddle-finance': saddleFinance,
   'saddle-finance-v2': saddleFinanceV2,
-  'saddle-vesdl': vesdl,
+  'saddle-sdl-balance-with-lock': sdlBalanceWithLock,
   'saddle-vesdl-voting-power': vesdlVotingPower,
   membership: membership,
   'lydia-gov-vault': lydiaGovVault,
